IM2: Fixes #4945

+++ provider-integration/im2/pkg/im/services/k8s/containers/00_module.go
@@ -307,22 +307,29 @@ func terminate(request ctrl.JobTerminateRequest) error {
 	// NOTE(Dan): Helper resources (e.g. Service) have a owner reference on the pod. For this reason, we do not need to
 	// delete this directly.
 	for rank := 0; rank < request.Job.Specification.Replicas; rank++ {
-		ctx, cancel := context.WithDeadline(context.Background(), time.Now().Add(10*time.Second))
 		podName := idAndRankToPodName(request.Job.Id, rank)
+		pod, ok := shared.JobPods.Retrieve(podName)
 
-		// NOTE(Dan): JobUpdateBatch and monitoring logic will aggressively get rid of pods that don't belong in
-		// the namespace and as such we don't have to worry about failures here.
-		_ = K8sClient.CoreV1().Pods(Namespace).Delete(ctx, podName, meta.DeleteOptions{
-			GracePeriodSeconds: util.Pointer[int64](1),
-		})
+		// NOTE(Dan): Do not waste time on pods that we know are not in the system. The K8s API is really slow and this
+		// can waste a lot of time if someone attempts to schedule a 1 million node job (which we cannot possibly host
+		// anyway).
+		if ok && pod.DeletionTimestamp == nil {
+			ctx, cancel := context.WithDeadline(context.Background(), time.Now().Add(10*time.Second))
+			// NOTE(Dan): JobUpdateBatch and monitoring logic will aggressively get rid of pods that don't belong in
+			// the namespace and as such we don't have to worry about failures here.
+			_ = K8sClient.CoreV1().Pods(Namespace).Delete(ctx, podName, meta.DeleteOptions{
+				GracePeriodSeconds: util.Pointer[int64](1),
+			})
 
-		cancel()
+			cancel()
+		}
 	}
 
 	// Unbinding IP and port assignments
 	// -----------------------------------------------------------------------------------------------------------------
 	ctrl.UnbindIpsFromJob(request.Job)
 	shared.ClearAssignedSshPort(request.Job)
+	shared.RemoveFromQueue(request.Job.Id)
 
 	// Cleaning up mount dirs
 	// -----------------------------------------------------------------------------------------------------------------
